tweak to how we calculate ec.hat
I've made a small tweak to the method with which we calculate ec.hat.
Originally, we just used q (the quantiles) in the denominator of
theta_ij (q). This led to the ratio log(Q.ij) / log(q) changing when it
actually should not have changed. Now, we're using log(Q.ij) /
log(q-hat) where q-hat is the estimated quantile from the data.